Update modal.js
I commented out the event listener for the overlay, which is supposed to make it so that you can close the modal when you click outside the modal and onto the overlay, because it is not working correctly. If I have time, I will try to get this working, but if I don't have time, then it is okay if we remove this feature since it is not a necessity for the rest of the modal functionality. The other changes that appear here are from my VSCode Prettier extension automatically changing the '' to "".