Update style.css
I deleted the z-index here and fixed it using Tailwind classes in the HTML. I also added bottom and right to the modal so that it covers the entire page and is visible when scrolling. (Before this fix, the overlay was appear to cover only to the height of the view port, so if you scrolled it was not there.) I also added a couple comments to the CSS to show what everything in here is for.